add course-sync service; clean up courseDo and schulconnex sync service; update tests + index
