+++ apps/server/src/modules/learnroom/service/index.ts
@@ -2,9 +2,10 @@ export * from './board-copy.service';
 export * from './common-cartridge-export.service';
 export * from './common-cartridge-import.service';
 export * from './course-copy.service';
-export * from './course.service';
 export { CourseDoService } from './course-do.service';
+export * from './course-rooms.service';
+export { CourseSyncService } from './course-sync.service';
+export * from './course.service';
 export * from './coursegroup.service';
 export * from './dashboard.service';
-export * from './course-rooms.service';
 export { GroupDeletedHandlerService } from './group-deleted-handler.service';
