+++ apps/server/src/modules/learnroom/uc/course-sync.uc.spec.ts
@@ -1,11 +1,11 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { AuthorizationContextBuilder, AuthorizationService } from '@modules/authorization';
 import { GroupService } from '@modules/group';
-import { RoleService } from '@modules/role';
 import { Test, TestingModule } from '@nestjs/testing';
-import { Permission, RoleName } from '@shared/domain/interface';
-import { groupFactory, roleDtoFactory, setupEntities, userFactory } from '@shared/testing';
+import { Permission } from '@shared/domain/interface';
+import { groupFactory, setupEntities, userFactory } from '@shared/testing';
 import { CourseDoService } from '../service';
+import { CourseSyncService } from '../service/course-sync.service';
 import { courseFactory } from '../testing';
 import { CourseSyncUc } from './course-sync.uc';
 
@@ -16,7 +16,7 @@ describe(CourseSyncUc.name, () => {
 	let authorizationService: DeepMocked<AuthorizationService>;
 	let courseService: DeepMocked<CourseDoService>;
 	let groupService: DeepMocked<GroupService>;
-	let roleService: DeepMocked<RoleService>;
+	let courseSyncService: DeepMocked<CourseSyncService>;
 
 	beforeAll(async () => {
 		module = await Test.createTestingModule({
@@ -31,12 +31,12 @@ describe(CourseSyncUc.name, () => {
 					useValue: createMock<CourseDoService>(),
 				},
 				{
-					provide: GroupService,
-					useValue: createMock<GroupService>(),
+					provide: CourseSyncService,
+					useValue: createMock<CourseSyncService>(),
 				},
 				{
-					provide: RoleService,
-					useValue: createMock<RoleService>(),
+					provide: GroupService,
+					useValue: createMock<GroupService>(),
 				},
 			],
 		}).compile();
@@ -45,7 +45,8 @@ describe(CourseSyncUc.name, () => {
 		authorizationService = module.get(AuthorizationService);
 		courseService = module.get(CourseDoService);
 		groupService = module.get(GroupService);
-		roleService = module.get(RoleService);
+		courseSyncService = module.get(CourseSyncService);
+
 		await setupEntities();
 	});
 
@@ -89,7 +90,7 @@ describe(CourseSyncUc.name, () => {
 
 				await uc.stopSynchronization(user.id, course.id);
 
-				expect(courseService.stopSynchronization).toHaveBeenCalledWith(course);
+				expect(courseSyncService.stopSynchronization).toHaveBeenCalledWith(course);
 			});
 		});
 	});
@@ -100,30 +101,15 @@ describe(CourseSyncUc.name, () => {
 				const user = userFactory.buildWithId();
 				const course = courseFactory.build();
 				const group = groupFactory.build();
-				const studentRole = roleDtoFactory.build({ id: 'student-role-id' });
-				const teacherRole = roleDtoFactory.build({ id: 'teacher-role-id' });
-
-				group.users = [
-					{ roleId: 'student-role-id', userId: 'student-user-id' },
-					{ roleId: 'teacher-role-id', userId: 'teacher-user-id' },
-				];
-
-				const students = group.users.filter((groupUser) => groupUser.roleId === studentRole.id);
-				const teachers = group.users.filter((groupUser) => groupUser.roleId === teacherRole.id);
 
 				courseService.findById.mockResolvedValueOnce(course);
 				groupService.findById.mockResolvedValueOnce(group);
 				authorizationService.getUserWithPermissions.mockResolvedValueOnce(user);
-				roleService.findByName.mockResolvedValueOnce(studentRole).mockResolvedValueOnce(teacherRole);
 
 				return {
 					user,
 					course,
 					group,
-					studentRole,
-					teacherRole,
-					students,
-					teachers,
 				};
 			};
 
@@ -132,8 +118,6 @@ describe(CourseSyncUc.name, () => {
 
 				await uc.startSynchronization(user.id, course.id, group.id);
 
-				expect(courseService.findById).toHaveBeenCalledWith(course.id);
-				expect(groupService.findById).toHaveBeenCalledWith(group.id);
 				expect(authorizationService.checkPermission).toHaveBeenCalledWith(
 					user,
 					course,
@@ -141,18 +125,28 @@ describe(CourseSyncUc.name, () => {
 				);
 			});
 
-			it('should start the synchronization with correct roles', async () => {
-				const { user, course, group, students, teachers } = setup();
+			it('should call course do service with the correct course id', async () => {
+				const { user, course, group } = setup();
 
 				await uc.startSynchronization(user.id, course.id, group.id);
 
 				expect(courseService.findById).toHaveBeenCalledWith(course.id);
+			});
+
+			it('should call group service with the correct group id', async () => {
+				const { user, course, group } = setup();
+
+				await uc.startSynchronization(user.id, course.id, group.id);
+
 				expect(groupService.findById).toHaveBeenCalledWith(group.id);
+			});
 
-				expect(roleService.findByName).toHaveBeenCalledWith(RoleName.STUDENT);
-				expect(roleService.findByName).toHaveBeenCalledWith(RoleName.TEACHER);
+			it('should start the synchronization', async () => {
+				const { user, course, group } = setup();
 
-				expect(courseService.startSynchronization).toHaveBeenCalledWith(course, group, students, teachers);
+				await uc.startSynchronization(user.id, course.id, group.id);
+
+				expect(courseSyncService.startSynchronization).toHaveBeenCalledWith(course, group);
 			});
 		});
 	});
