+++ apps/server/src/modules/provisioning/strategy/oidc/service/schulconnex-course-sync.service.spec.ts
@@ -1,39 +1,28 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
-import { ObjectId } from '@mikro-orm/mongodb';
-import { Group, GroupUser } from '@modules/group';
-import { CourseDoService } from '@modules/learnroom';
-import { Course } from '@modules/learnroom/domain';
-import { courseFactory } from '@modules/learnroom/testing';
-import { RoleDto, RoleService } from '@modules/role';
+import { Group } from '@modules/group';
+import { CourseSyncService } from '@modules/learnroom';
 import { Test, TestingModule } from '@nestjs/testing';
-import { groupFactory, roleDtoFactory } from '@shared/testing';
+import { groupFactory } from '@shared/testing';
 import { SchulconnexCourseSyncService } from './schulconnex-course-sync.service';
 
 describe(SchulconnexCourseSyncService.name, () => {
 	let module: TestingModule;
 	let service: SchulconnexCourseSyncService;
-
-	let courseService: DeepMocked<CourseDoService>;
-	let roleService: DeepMocked<RoleService>;
+	let courseSyncService: DeepMocked<CourseSyncService>;
 
 	beforeAll(async () => {
 		module = await Test.createTestingModule({
 			providers: [
 				SchulconnexCourseSyncService,
 				{
-					provide: CourseDoService,
-					useValue: createMock<CourseDoService>(),
-				},
-				{
-					provide: RoleService,
-					useValue: createMock<RoleService>(),
+					provide: CourseSyncService,
+					useValue: createMock<CourseSyncService>(),
 				},
 			],
 		}).compile();
 
 		service = module.get(SchulconnexCourseSyncService);
-		courseService = module.get(CourseDoService);
-		roleService = module.get(RoleService);
+		courseSyncService = module.get(CourseSyncService);
 	});
 
 	afterAll(async () => {
@@ -47,184 +36,39 @@ describe(SchulconnexCourseSyncService.name, () => {
 	describe('synchronizeCourseWithGroup', () => {
 		describe('when synchronizing with a new group', () => {
 			const setup = () => {
-				const course: Course = courseFactory.build();
-				const studentId: string = new ObjectId().toHexString();
-				const teacherId: string = new ObjectId().toHexString();
-				const studentRoleId: string = new ObjectId().toHexString();
-				const teacherRoleId: string = new ObjectId().toHexString();
-				const studentRole: RoleDto = roleDtoFactory.build({ id: studentRoleId });
-				const teacherRole: RoleDto = roleDtoFactory.build({ id: teacherRoleId });
-				const newGroup: Group = groupFactory.build({
-					users: [
-						{
-							userId: studentId,
-							roleId: studentRoleId,
-						},
-						{
-							userId: teacherId,
-							roleId: teacherRoleId,
-						},
-					],
-				});
-
-				courseService.findBySyncedGroup.mockResolvedValueOnce([new Course(course.getProps())]);
-				roleService.findByName.mockResolvedValueOnce(studentRole);
-				roleService.findByName.mockResolvedValueOnce(teacherRole);
+				const newGroup: Group = groupFactory.build();
 
 				return {
-					course,
 					newGroup,
-					studentId,
-					teacherId,
 				};
 			};
 
 			it('should synchronize with the group', async () => {
-				const { course, newGroup, studentId, teacherId } = setup();
+				const { newGroup } = setup();
 
 				await service.synchronizeCourseWithGroup(newGroup);
 
-				expect(courseService.saveAll).toHaveBeenCalledWith<[Course[]]>([
-					new Course({
-						...course.getProps(),
-						name: newGroup.name,
-						startDate: newGroup.validPeriod?.from,
-						untilDate: newGroup.validPeriod?.until,
-						studentIds: [studentId],
-						teacherIds: [teacherId],
-					}),
-				]);
+				expect(courseSyncService.synchronizeCourseWithGroup).toHaveBeenCalledWith(newGroup, undefined);
 			});
 		});
 
-		describe('when the course name is the same as the old group name', () => {
+		describe('when synchronizing with a new group and an old group', () => {
 			const setup = () => {
-				const course: Course = courseFactory.build({ name: 'Course Name' });
-				const studentRole: RoleDto = roleDtoFactory.build();
-				const teacherRole: RoleDto = roleDtoFactory.build();
-				const oldGroup: Group = groupFactory.build({ name: 'Course Name' });
-				const newGroup: Group = groupFactory.build({
-					name: 'New Group Name',
-					users: [],
-				});
-
-				courseService.findBySyncedGroup.mockResolvedValueOnce([new Course(course.getProps())]);
-				roleService.findByName.mockResolvedValueOnce(studentRole);
-				roleService.findByName.mockResolvedValueOnce(teacherRole);
+				const newGroup: Group = groupFactory.build();
+				const oldGroup: Group = groupFactory.build();
 
 				return {
-					course,
 					newGroup,
 					oldGroup,
 				};
 			};
 
-			it('should synchronize the group name', async () => {
-				const { course, newGroup, oldGroup } = setup();
+			it('should synchronize with the group', async () => {
+				const { newGroup, oldGroup } = setup();
 
 				await service.synchronizeCourseWithGroup(newGroup, oldGroup);
 
-				expect(courseService.saveAll).toHaveBeenCalledWith<[Course[]]>([
-					new Course({
-						...course.getProps(),
-						name: newGroup.name,
-						startDate: newGroup.validPeriod?.from,
-						untilDate: newGroup.validPeriod?.until,
-						studentIds: [],
-						teacherIds: [],
-					}),
-				]);
-			});
-		});
-
-		describe('when the course name is different from the old group name', () => {
-			const setup = () => {
-				const course: Course = courseFactory.build({ name: 'Custom Course Name' });
-				const studentRole: RoleDto = roleDtoFactory.build();
-				const teacherRole: RoleDto = roleDtoFactory.build();
-				const oldGroup: Group = groupFactory.build({ name: 'Course Name' });
-				const newGroup: Group = groupFactory.build({
-					name: 'New Group Name',
-					users: [],
-				});
-
-				courseService.findBySyncedGroup.mockResolvedValueOnce([new Course(course.getProps())]);
-				roleService.findByName.mockResolvedValueOnce(studentRole);
-				roleService.findByName.mockResolvedValueOnce(teacherRole);
-
-				return {
-					course,
-					newGroup,
-					oldGroup,
-				};
-			};
-
-			it('should keep the old course name', async () => {
-				const { course, newGroup, oldGroup } = setup();
-
-				await service.synchronizeCourseWithGroup(newGroup, oldGroup);
-
-				expect(courseService.saveAll).toHaveBeenCalledWith<[Course[]]>([
-					new Course({
-						...course.getProps(),
-						name: course.name,
-						startDate: newGroup.validPeriod?.from,
-						untilDate: newGroup.validPeriod?.until,
-						studentIds: [],
-						teacherIds: [],
-					}),
-				]);
-			});
-		});
-
-		describe('when the last teacher gets removed from a synced group', () => {
-			const setup = () => {
-				const studentUserId = new ObjectId().toHexString();
-				const teacherUserId = new ObjectId().toHexString();
-				const course: Course = courseFactory.build({
-					teacherIds: [teacherUserId],
-					studentIds: [studentUserId],
-					syncedWithGroup: new ObjectId().toHexString(),
-				});
-				const studentRoleId: string = new ObjectId().toHexString();
-				const studentRole: RoleDto = roleDtoFactory.build({ id: studentRoleId });
-				const teacherRole: RoleDto = roleDtoFactory.build();
-				const newGroup: Group = groupFactory.build({
-					users: [
-						new GroupUser({
-							userId: studentUserId,
-							roleId: studentRoleId,
-						}),
-					],
-				});
-
-				courseService.findBySyncedGroup.mockResolvedValueOnce([new Course(course.getProps())]);
-				roleService.findByName.mockResolvedValueOnce(studentRole);
-				roleService.findByName.mockResolvedValueOnce(teacherRole);
-
-				return {
-					course,
-					newGroup,
-					teacherUserId,
-				};
-			};
-
-			it('should keep the last teacher, remove all students', async () => {
-				const { course, newGroup, teacherUserId } = setup();
-
-				await service.synchronizeCourseWithGroup(newGroup, newGroup);
-
-				expect(courseService.saveAll).toHaveBeenCalledWith<[Course[]]>([
-					new Course({
-						...course.getProps(),
-						name: course.name,
-						startDate: newGroup.validPeriod?.from,
-						untilDate: newGroup.validPeriod?.until,
-						studentIds: [],
-						teacherIds: [teacherUserId],
-						syncedWithGroup: course.syncedWithGroup,
-					}),
-				]);
+				expect(courseSyncService.synchronizeCourseWithGroup).toHaveBeenCalledWith(newGroup, oldGroup);
 			});
 		});
 	});
