+++ apps/server/src/modules/provisioning/strategy/oidc/service/schulconnex-course-sync.service.ts
@@ -1,46 +1,12 @@
-import { Group, GroupUser } from '@modules/group';
-import { CourseDoService } from '@modules/learnroom';
-import { RoleDto, RoleService } from '@modules/role';
+import { Group } from '@modules/group';
+import { CourseSyncService } from '@modules/learnroom';
 import { Injectable } from '@nestjs/common';
-import { RoleName } from '@shared/domain/interface';
-import { EntityId } from '@shared/domain/types';
-import { Course } from '@src/modules/learnroom/domain';
 
 @Injectable()
 export class SchulconnexCourseSyncService {
-	constructor(private readonly courseService: CourseDoService, private readonly roleService: RoleService) {}
+	constructor(private readonly courseSyncService: CourseSyncService) {}
 
 	async synchronizeCourseWithGroup(newGroup: Group, oldGroup?: Group): Promise<void> {
-		const courses: Course[] = await this.courseService.findBySyncedGroup(newGroup);
-
-		if (courses.length) {
-			const studentRole: RoleDto = await this.roleService.findByName(RoleName.STUDENT);
-			const teacherRole: RoleDto = await this.roleService.findByName(RoleName.TEACHER);
-
-			courses.forEach((course: Course): void => {
-				if (!oldGroup || oldGroup.name === course.name) {
-					course.name = newGroup.name;
-				}
-
-				course.startDate = newGroup.validPeriod?.from;
-				course.untilDate = newGroup.validPeriod?.until;
-
-				const students: GroupUser[] = newGroup.users.filter(
-					(user: GroupUser): boolean => user.roleId === studentRole.id
-				);
-				const teachers: GroupUser[] = newGroup.users.filter(
-					(user: GroupUser): boolean => user.roleId === teacherRole.id
-				);
-
-				if (teachers.length >= 1) {
-					course.students = students.map((user: GroupUser): EntityId => user.userId);
-					course.teachers = teachers.map((user: GroupUser): EntityId => user.userId);
-				} else {
-					course.students = [];
-				}
-			});
-
-			await this.courseService.saveAll(courses);
-		}
+		await this.courseSyncService.synchronizeCourseWithGroup(newGroup, oldGroup);
 	}
 }
